requirements_test.txt
-r requirements.txt

pytest==6.2.5
pytest-mock==3.7.0
arrow==1.2.2
factory-boy==3.2.1
pytest-factoryboy==2.1.0
pytest-cov==3.0.0
